requirements.txt
# Core dependencies
Flask==2.3.3
PyYAML==6.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Testing
pytest>=7.0.0

# For documentation
sphinx>=7.0.0 